+++ MASsenger.Application/Dtos/Create/MessageCreateDto.cs
@@ -2,8 +2,8 @@
 {
     public class MessageCreateDto
     {
-        public ulong SenderID { set; get; }
-        public ulong DestinationID { set; get; }
+        public Int32 SenderID { set; get; }
+        public Int32 DestinationID { set; get; }
         public string? Text { set; get; }
         public DateTime? SentTime { set; get; }
     }
